+++ src/main/java/net/witcher_rpg/item/armor/Armors.java
@@ -12,6 +12,7 @@ import net.minecraft.sound.SoundEvents;
 import net.minecraft.util.Identifier;
 import net.spell_engine.api.config.ArmorSetConfig;
 import net.spell_engine.api.config.AttributeModifier;
+import net.spell_engine.api.item.Equipment;
 import net.spell_power.api.SpellPowerMechanics;
 import net.witcher_rpg.item.WitcherGroup;
 import net.spell_engine.api.item.armor.Armor;
@@ -58,13 +59,16 @@ public class Armors {
         return Registry.registerReference(Registries.ARMOR_MATERIAL, Identifier.of(MOD_ID, name), material);
     }
     public static final ArrayList<Armor.Entry> entries = new ArrayList<>();
-    private static Armor.Entry create(RegistryEntry<ArmorMaterial> material, Identifier id, int durability, Armor.Set.ItemFactory factory, ArmorSetConfig defaults) {
+    private static Armor.Entry create(RegistryEntry<ArmorMaterial> material, Identifier id, int durability,
+                                      Armor.Set.ItemFactory factory, ArmorSetConfig defaults, int tier) {
         var entry = Armor.Entry.create(
                 material,
                 id,
                 durability,
                 factory,
-                defaults);
+                defaults,
+                Equipment.LootProperties.of(tier)
+        );
         entries.add(entry);
         return entry;
     }
@@ -215,7 +219,7 @@ public class Armors {
                                             AttributeModifier.multiply(SIGN_INTENSITY,witcherSign),
                                             AttributeModifier.multiply(ADRENALINE,witcherAdrenaline),
                                             AttributeModifier.multiply(ATTACK_DAMAGE,witcherAttackDamage)
-                                    )))).armorSet();
+                                    ))),1).armorSet();
 
     //FELINE SCHOOL ARMOR
     public static final Armor.Set felineSchoolArmorSet =
@@ -241,7 +245,7 @@ public class Armors {
                                     .addAll(List.of(
                                             AttributeModifier.multiply(ADRENALINE,felineAdrenaline),
                                             AttributeModifier.multiply(ATTACK_SPEED,felineAttackSpeed)
-                                    )))).armorSet();
+                                    ))),1).armorSet();
 
     public static final Armor.Set enhancedFelineSchoolArmorSet =
             create(
@@ -270,7 +274,7 @@ public class Armors {
                                             AttributeModifier.multiply(ADRENALINE,felineAdrenalineT2),
                                             AttributeModifier.multiply(ATTACK_SPEED,felineAttackSpeedT2),
                                             AttributeModifier.multiply(ATTACK_DAMAGE,felineAttackDamageT2)
-                                    )))).armorSet();
+                                    ))),2).armorSet();
 
     public static final Armor.Set superiorFelineSchoolArmorSet =
             create(
@@ -303,7 +307,7 @@ public class Armors {
                                             AttributeModifier.multiply(ATTACK_SPEED,felineAttackSpeedT3),
                                             AttributeModifier.multiply(ATTACK_DAMAGE,felineAttackDamageT3),
                                             AttributeModifier.multiply(COMBATROLL_RECHARGE,felineRollRechargeT3)
-                                    )))).armorSet();
+                                    ))),3).armorSet();
     //GRIFFIN SCHOOL ARMOR
     public static final Armor.Set griffinArmorSet =
             create(
@@ -328,7 +332,7 @@ public class Armors {
                                     .addAll(List.of(
                                             AttributeModifier.multiply(SIGN_INTENSITY,griffinSign),
                                             AttributeModifier.multiply(ADRENALINE,griffinAdrenaline)
-                                    )))).armorSet();
+                                    ))),1).armorSet();
 
     public static final Armor.Set enhancedGriffinArmorSet =
             create(
@@ -357,7 +361,7 @@ public class Armors {
                                             AttributeModifier.multiply(SIGN_INTENSITY,griffinSignT2),
                                             AttributeModifier.multiply(ADRENALINE,griffinAdrenalineT2),
                                             AttributeModifier.multiply(SpellPowerMechanics.HASTE.id,griffinHasteT2)
-                                    )))).armorSet();
+                                    ))),2).armorSet();
 
     public static final Armor.Set superiorGriffinArmorSet =
             create(
@@ -386,7 +390,7 @@ public class Armors {
                                             AttributeModifier.multiply(SIGN_INTENSITY,griffinSignT3),
                                             AttributeModifier.multiply(ADRENALINE,griffinAdrenalineT3),
                                             AttributeModifier.multiply(SpellPowerMechanics.HASTE.id,griffinHasteT3)
-                                    )))).armorSet();
+                                    ))),3).armorSet();
 
     //WOLVEN SCHOOL ARMOR
     public static final Armor.Set wolvenArmorSet =
@@ -412,7 +416,7 @@ public class Armors {
                                     .addAll(List.of(
                                             AttributeModifier.multiply(SIGN_INTENSITY,wolvenSign),
                                             AttributeModifier.multiply(ATTACK_DAMAGE,wolvenAttackDamage)
-                                    )))).armorSet();
+                                    ))),1).armorSet();
     public static final Armor.Set enhancedWolvenArmorSet =
             create(
                     material_enhanced_wolven, Identifier.of(MOD_ID, "enhanced_wolven"), 20, WolfSchoolArmor::new,
@@ -440,7 +444,7 @@ public class Armors {
                                             AttributeModifier.multiply(SIGN_INTENSITY,wolvenSignT2),
                                             AttributeModifier.multiply(ATTACK_DAMAGE,wolvenAttackDamageT2),
                                             AttributeModifier.multiply(ADRENALINE,wolvenAdrenalineT2)
-                                    )))).armorSet();
+                                    ))),2).armorSet();
     public static final Armor.Set superiorWolvenArmorSet =
             create(
                     material_superior_wolven, Identifier.of(MOD_ID, "superior_wolven"), 25, WolfSchoolArmor::new,
@@ -468,7 +472,7 @@ public class Armors {
                                             AttributeModifier.multiply(SIGN_INTENSITY,wolvenSignT3),
                                             AttributeModifier.multiply(ATTACK_DAMAGE,wolvenAttackDamageT3),
                                             AttributeModifier.multiply(ADRENALINE,wolvenAdrenalineT3)
-                                    )))).armorSet();
+                                    ))),3).armorSet();
 
     //URSINE SCHOOL ARMOR
     public static final Armor.Set ursineArmorSet =
@@ -494,7 +498,7 @@ public class Armors {
                                     .addAll(List.of(
                                             AttributeModifier.multiply(ADRENALINE,ursineAdrenaline),
                                             AttributeModifier.multiply(KNOCKBACK_RESISTANCE,ursineKnockBackResi)
-                                    )))).armorSet();
+                                    ))),1).armorSet();
     public static final Armor.Set enhancedUrsineArmorSet =
             create(
                     material_enhanced_ursine, Identifier.of(MOD_ID, "enhanced_ursine"), 20, BearSchoolArmor::new,
@@ -522,7 +526,7 @@ public class Armors {
                                             AttributeModifier.multiply(ADRENALINE,ursineAdrenalineT2),
                                             AttributeModifier.multiply(KNOCKBACK_RESISTANCE,ursineKnockBackResiT2),
                                             AttributeModifier.multiply(ATTACK_DAMAGE,ursineAttackDamageT2)
-                                    )))).armorSet();
+                                    ))),2).armorSet();
     public static final Armor.Set superiorUrsineArmorSet =
             create(
                     material_superior_ursine, Identifier.of(MOD_ID, "superior_ursine"), 25, BearSchoolArmor::new,
@@ -554,7 +558,7 @@ public class Armors {
                                             AttributeModifier.multiply(KNOCKBACK_RESISTANCE,ursineKnockBackResiT3),
                                             AttributeModifier.multiply(ATTACK_DAMAGE,ursineAttackDamageT3),
                                             AttributeModifier.bonus(ARMOR_TOUGHNESS,ursineArmorToughnessT3)
-                                    )))).armorSet();
+                                    ))),3).armorSet();
 
 
     public static void register(Map<String,  ArmorSetConfig> configs) {